+++ TIGEST/Functions/Frm_GestaoUsuarios.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+using TIGEST.Models;
+
+namespace TIGEST.Functions
+{
+    public partial class Frm_GestaoUsuarios: Form
+    {
+        public Frm_GestaoUsuarios()
+        {
+            InitializeComponent();
+        }
+    }
+}
