--- TIGEST/Model/Dados.cs
@@ -1,9 +1,0 @@
-﻿using System.Data.SqlClient;
-
-namespace TIGEST.Model
-{
-    internal class Dados
-    {
-
-    }
-}
